FossilOrigin-Name: 6759efcc813e66cf937413b4a162eae6bfed4d4b1322721202296d9c9755711d

+++ src/delta.c
@@ -604,7 +604,7 @@ int delta_apply(
           /* ERROR: copy exceeds output file size */
           return -1;
         }
-        if( (int)(ofst+cnt) > lenSrc ){
+        if( (u64)ofst+(u64)cnt > (u64)lenSrc ){
           /* ERROR: copy extends past end of input */
           return -1;
         }
